short_scoring
Experimental branch to see if representing the per IP topic probabilities in 16 bits each saves us significant time and memory.

If this branch doesn't show the desired improvements... the only thing to do would be to go into spark lda and change it's representation of the probabilities, which is more bother than it's probably worth.